// src/types/trade.ts
export interface Trade {
  aptDong: string;
  aptNm: string;
  buildYear: number;
  buyerGbn: string;
  dealAmount: number;   // 🔥 숫자로 변환
  dealDay: number;
  dealMonth: number;
  dealYear: number;
  excluUseAr: number;
  floor: number;
  jibun: string;
  sggCd: string;
  umdNm: string;
}